+++ libmat/srcs/ft_mat_vec_mult.c
@@ -12,7 +12,7 @@
 
 #include "libmat.h"
 
-void	ft_load_vec(t_point3d vec, float vec_as_arr[4])
+void	ft_load_vec(t_point3d vec, double vec_as_arr[4])
 {
 	vec_as_arr[0] = vec.x;
 	vec_as_arr[1] = vec.y;
@@ -20,7 +20,7 @@ void	ft_load_vec(t_point3d vec, float vec_as_arr[4])
 	vec_as_arr[3] = 1.0;
 }
 
-void	ft_init_vec(float vec_as_arr[3])
+void	ft_init_vec(double vec_as_arr[3])
 {
 	vec_as_arr[0] = 0;
 	vec_as_arr[1] = 0;
@@ -29,8 +29,8 @@ void	ft_init_vec(float vec_as_arr[3])
 
 t_point3d	ft_point3d_mat4_mult(t_point3d vec, t_mat4 mat)
 {
-	float vec_as_arr[4];
-	float res[3];
+	double vec_as_arr[4];
+	double res[3];
 	int i;
 	int j;
 	int k;
