+++ libmat/srcs/ft_normalize.c
@@ -12,15 +12,15 @@
 
 #include "libmat.h"
 
-float	ft_point3d_get_norm(t_point3d vec)
+double	ft_point3d_get_norm(t_point3d vec)
 {
 	return (sqrt(vec.x * vec.x + vec.y * vec.y + vec.z * vec.z));
 }
 
 void	ft_point3d_normalize(t_point3d *vec)
 {
-	float norm;
-	float invnorm;
+	double norm;
+	double invnorm;
 
 	norm = ft_point3d_get_norm(*vec);
 	if (norm != 0)
